testing/test_requirements.txt
nltk==3.8.1
pytest-cov==4.1.0
pandas==1.3.5